+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes in PAPPL
 Changes in v1.3b1
 -----------------
 
+- Added timer APIs to manage periodic tasks (Issue #208)
 - Added debug logging for device management.
 - Fixed a device race condition with job processing.
 - Fixed a potential value overflow when reading SNMP OIDs (Issue #210)